resolve on comments
