+++ tests/loops/test_evaluation_loop.py
@@ -13,9 +13,8 @@
 # limitations under the License.
 from unittest import mock
 
-import pytest
 import torch
-from torch.utils.data.dataloader import _MultiProcessingDataLoaderIter, DataLoader
+from torch.utils.data.dataloader import DataLoader
 
 from pytorch_lightning import Trainer
 from pytorch_lightning.loops import EvaluationEpochLoop
@@ -131,33 +130,3 @@ def test_evaluation_loop_doesnt_store_outputs_if_epoch_end_not_overridden(tmpdir
     trainer.test_loop.connect(TestLoop())
     trainer.test(model)
     assert did_assert
-
-
-@RunIf(min_torch="1.8.0")
-@pytest.mark.parametrize("persistent_workers", (True, False))
-def test_evaluation_workers_are_shutdown(tmpdir, persistent_workers):
-    # `num_workers == 1` uses `_MultiProcessingDataLoaderIter`
-    # `persistent_workers` makes sure `self._iterator` gets set on the `DataLoader` instance
-    has_shutdown_workers = False
-
-    class _TestMultiProcessingDataLoaderIter(_MultiProcessingDataLoaderIter):
-        def _shutdown_workers(self):
-            nonlocal has_shutdown_workers
-            has_shutdown_workers = True
-            super()._shutdown_workers()
-
-    class TestDataLoader(DataLoader):
-        def _get_iterator(self):
-            if self.num_workers == 0:
-                return super()._get_iterator()
-            else:
-                self.check_worker_number_rationality()
-                return _TestMultiProcessingDataLoaderIter(self)
-
-    train_dataloader = DataLoader(RandomDataset(32, 64), num_workers=0)
-    val_dataloader = TestDataLoader(RandomDataset(32, 64), num_workers=1, persistent_workers=persistent_workers)
-    model = BoringModel()
-    trainer = Trainer(default_root_dir=tmpdir, limit_train_batches=2, limit_val_batches=2, max_epochs=2)
-    trainer.fit(model, train_dataloader, val_dataloader)
-    assert has_shutdown_workers
-    assert val_dataloader._iterator is None
